Added the control task to receive and enable the heater or humidifier.
Added queues to communicate between the callback timer, the serial task and the control task.

The callback timer send to the control task the sensor data.

The serial task send to the control task the setpoint of temperature and humidity.

The control task compare the values with the setpoint and enable or disable the heater or humidifier.